Publish Service clleaned up and extras
- Microservices now wait for DB to launch and ensure that the service is healthy.
- Removed extra semi-colon
- You can now only publish your own books and delete your own books. It automatically grabs your publisherId
- Updated User and Book models.
- JWTFilter now knows the current username based on last used local JWT Token.
- Updated Order-Service application properties.
- Added Rest Template for Publishing-Service.
- Removed Temporary User Controller and UserService in Publishing-Service.

+++ MicroServices/PublishingService/src/main/java/cs4337/group6/PublishingService/Services/PublishingService.java
@@ -4,9 +4,11 @@ import cs4337.group6.PublishingService.Models.Book;
 import cs4337.group6.PublishingService.Models.User;
 import cs4337.group6.PublishingService.Repositories.IBookRepository;
 import cs4337.group6.PublishingService.Repositories.IUserRepository;
+import org.hibernate.Hibernate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -22,22 +24,40 @@ public class PublishingService
      * @param bookToAdd The book to add.
      * @return The book if added or null if not.
      */
-    public Book PublishBook(Book bookToAdd, Integer publisherId)
+    public Book PublishBook(Book bookToAdd, String username)
     {
-        User publisher = userRepo.findById(publisherId).orElseThrow(() -> new RuntimeException("Publisher not found"));
+        User publisher = userRepo.findByUsername(username);
 
+        if (publisher == null) {
+        throw new IllegalArgumentException("No user found with username: " + username);
+        }
+        //System.out.println(publisher);
         bookToAdd.setPublisher(publisher);
+
         return repository.save(bookToAdd);
     }
 
-    public void RemoveBook(Integer bookId)
-    {
+    public void RemoveBook(Integer bookId, String username) throws Exception {
+        User currentUser = userRepo.findByUsername(username);
         Book book = repository.findById(bookId).orElseThrow(() -> new RuntimeException("Book Not Found"));
-        repository.delete(book);
+        if(book.getPublisher().equals(currentUser))
+        {
+            repository.delete(book);
+        }
+        else
+        {
+            throw new Exception("You are not the publisher of this book!");
+        }
     }
 
     public List<Book> GetAllPublishedBooks()
     {
-        return repository.findAll();
+        List<Book> books = repository.findAll();
+        books.forEach(book -> Hibernate.initialize(book.getBuyers()));
+        books.forEach(book -> {
+            book.setBuyers(Collections.unmodifiableSet(book.getBuyers()));
+        });
+
+        return Collections.unmodifiableList(books);
     }
 }
